Alt click + terminal button to split (#123531)
