+++ src/vs/platform/actions/browser/dropdownWithPrimaryActionViewItem.ts
@@ -11,6 +11,10 @@ import * as DOM from 'vs/base/browser/dom';
 import { StandardKeyboardEvent } from 'vs/base/browser/keyboardEvent';
 import { KeyCode } from 'vs/base/common/keyCodes';
 import { dispose, IDisposable } from 'vs/base/common/lifecycle';
+import { MenuItemAction } from 'vs/platform/actions/common/actions';
+import { MenuEntryActionViewItem } from 'vs/platform/actions/browser/menuEntryActionViewItem';
+import { IKeybindingService } from 'vs/platform/keybinding/common/keybinding';
+import { INotificationService } from 'vs/platform/notification/common/notification';
 
 export class DropdownWithPrimaryActionViewItem extends BaseActionViewItem {
 	private _primaryAction: ActionViewItem;
@@ -20,18 +24,16 @@ export class DropdownWithPrimaryActionViewItem extends BaseActionViewItem {
 	private toDispose: IDisposable[];
 
 	constructor(
-		primaryAction: IAction,
+		primaryAction: MenuItemAction,
 		dropdownAction: IAction,
 		dropdownMenuActions: IAction[],
-		_className: string,
+		className: string,
 		private readonly _contextMenuProvider: IContextMenuProvider,
-		dropdownIcon?: string
+		_keybindingService: IKeybindingService,
+		_notificationService: INotificationService
 	) {
 		super(null, primaryAction);
-		this._primaryAction = new ActionViewItem(undefined, primaryAction, {
-			icon: true,
-			label: false
-		});
+		this._primaryAction = new MenuEntryActionViewItem(primaryAction, _keybindingService, _notificationService);
 		this._dropdown = new DropdownMenuActionViewItem(dropdownAction, dropdownMenuActions, this._contextMenuProvider, {
 			menuAsChild: true
 		});
@@ -46,7 +48,6 @@ export class DropdownWithPrimaryActionViewItem extends BaseActionViewItem {
 		this._primaryAction.render(DOM.append(this._container, primaryContainer));
 		this._dropdownContainer = DOM.$('.dropdown-action-container');
 		this._dropdown.render(DOM.append(this._container, this._dropdownContainer));
-
 		this.toDispose.push(DOM.addDisposableListener(primaryContainer, DOM.EventType.KEY_DOWN, (e: KeyboardEvent) => {
 			const event = new StandardKeyboardEvent(e);
 			if (event.equals(KeyCode.RightArrow)) {
